reimplemented generics for commands for advanced commands. standardized return types

+++ src/slogo/model/backEndInternal/commands/And.java
@@ -1,6 +1,6 @@
 package slogo.model.backEndInternal.commands;
 
-public class And implements Command<Integer> {
+public class And implements Command<Double> {
 
   Double value1;
   Double value2;
@@ -11,7 +11,7 @@ public class And implements Command<Integer> {
   }
 
   @Override
-  public double execute() {
-    return (value1 != 0 && value2 != 0 ) ? 1 : 0;
+  public Double execute() {
+    return (double) ((value1 != 0 && value2 != 0) ? 1 : 0);
   }
 }
