+++ src/slogo/model/backEndInternal/commands/ShowTurtle.java
@@ -2,7 +2,7 @@ package slogo.model.backEndInternal.commands;
 
 import slogo.model.backEndInternal.BackEndTurtle;
 
-public class ShowTurtle implements Command<Integer> {
+public class ShowTurtle implements Command<Double> {
 
   BackEndTurtle backEndTurtle;
 
@@ -11,8 +11,8 @@ public class ShowTurtle implements Command<Integer> {
   }
 
   @Override
-  public double execute() {
+  public Double execute() {
     backEndTurtle.toggleVisibility();
-    return 1;
+    return 1.0;
   }
 }
